feature: Restage changed files and fix only touched lines
