+++ .golangci.yml
@@ -9,36 +9,50 @@ linters-settings:
 linters:
   disable-all: true
   enable:
-   - bodyclose
-   - depguard
-   - dogsled
-   - dupl
-   - errcheck
-   - exhaustive
-   - exportloopref
-   - gci
-   - goconst
-   - gocyclo
-   - gocyclo
-   - godot
-   - godox
-   - gofmt
-   - gofumpt
-   - goimports
-   - gomnd
-   - goprintffuncname
-   - gosimple
-   - govet
-   - ineffassign
-   - misspell
-   - nakedret
-   - nestif
-   - noctx
-   - nolintlint
-   - revive
-   - staticcheck
-   - typecheck
-   - unconvert
-   - unparam
-   - unused
-   - whitespace
+    - asasalint
+    - asciicheck
+    - bidichk
+    - bodyclose
+    - containedctx
+    - contextcheck
+    - decorder
+    - depguard
+    - dogsled
+    - dupl
+    - dupword
+    - durationcheck
+    - errcheck
+    - errchkjson
+    - errname
+    - errorlint
+    - exhaustive
+    - exportloopref
+    - forbidigo
+    - gci
+    - gochecknoinits
+    - goconst
+    - gocritic
+    - gocyclo
+    - godot
+    - godox
+    - gofmt
+    - gofumpt
+    - goheader
+    - goimports
+    - gomnd
+    - goprintffuncname
+    - gosimple
+    - govet
+    - ineffassign
+    - misspell
+    - nakedret
+    - nestif
+    - noctx
+    - nolintlint
+    - revive
+    - staticcheck
+    - typecheck
+    - unconvert
+    - unparam
+    - unused
+    - whitespace
