+++ internal/lefthook/runner/runner.go
@@ -13,7 +13,6 @@ import (
 	"sync/atomic"
 
 	"github.com/spf13/afero"
-	"gopkg.in/alessio/shellescape.v1"
 
 	"github.com/evilmartians/lefthook/internal/config"
 	"github.com/evilmartians/lefthook/internal/git"
@@ -33,6 +32,7 @@ type Opts struct {
 	Fs           afero.Fs
 	Repo         *git.Repository
 	Hook         *config.Hook
+	HookName     string
 	GitArgs      []string
 	ResultChan   chan Result
 	SkipSettings log.SkipSettings
@@ -56,13 +56,13 @@ func NewRunner(opts Opts) *Runner {
 
 // RunAll runs scripts and commands.
 // LFS hook is executed at first if needed.
-func (r *Runner) RunAll(hookName string, sourceDirs []string) {
-	if err := r.runLFSHook(hookName); err != nil {
+func (r *Runner) RunAll(sourceDirs []string) {
+	if err := r.runLFSHook(r.HookName); err != nil {
 		log.Error(err)
 	}
 
 	if r.Hook.Skip != nil && r.Hook.DoSkip(r.Repo.State()) {
-		logSkip(hookName, "(SKIP BY HOOK SETTING)")
+		logSkip(r.HookName, "hook setting")
 		return
 	}
 
@@ -74,7 +74,7 @@ func (r *Runner) RunAll(hookName string, sourceDirs []string) {
 	scriptDirs := make([]string, len(sourceDirs))
 	for _, sourceDir := range sourceDirs {
 		scriptDirs = append(scriptDirs, filepath.Join(
-			sourceDir, hookName,
+			sourceDir, r.HookName,
 		))
 	}
 
@@ -115,12 +115,13 @@ func (r *Runner) runLFSHook(hookName string) error {
 		log.Debugf(
 			"[git-lfs] executing hook: git lfs %s %s", hookName, strings.Join(r.GitArgs, " "),
 		)
-		out, err := r.executor.RawExecute(
-			"git",
+		out := bytes.NewBuffer(make([]byte, 0))
+		err := r.executor.RawExecute(
 			append(
-				[]string{"lfs", hookName},
+				[]string{"git", "lfs", hookName},
 				r.GitArgs...,
-			)...,
+			),
+			out,
 		)
 
 		output := strings.Trim(out.String(), "\n")
@@ -169,12 +170,12 @@ func (r *Runner) runScripts(dir string) {
 	for _, file := range files {
 		script, ok := r.Hook.Scripts[file.Name()]
 		if !ok {
-			logSkip(file.Name(), "(SKIP BY NOT EXIST IN CONFIG)")
+			logSkip(file.Name(), "not specified in config file")
 			continue
 		}
 
 		if r.failed.Load() && r.Hook.Piped {
-			logSkip(file.Name(), "(SKIP BY BROKEN PIPE)")
+			logSkip(file.Name(), "broken pipe")
 			continue
 		}
 
@@ -201,7 +202,7 @@ func (r *Runner) runScripts(dir string) {
 	for _, file := range interactiveScripts {
 		script := r.Hook.Scripts[file.Name()]
 		if r.failed.Load() {
-			logSkip(file.Name(), "(SKIP INTERACTIVE BY FAILED)")
+			logSkip(file.Name(), "non-interactive scripts failed")
 			continue
 		}
 
@@ -212,38 +213,11 @@ func (r *Runner) runScripts(dir string) {
 }
 
 func (r *Runner) runScript(script *config.Script, path string, file os.FileInfo) {
-	if script.Skip != nil && script.DoSkip(r.Repo.State()) {
-		logSkip(file.Name(), "(SKIP BY SETTINGS)")
+	args, err := r.prepareScript(script, path, file)
+	if err != nil {
+		logSkip(file.Name(), err.Error())
 		return
 	}
-
-	if intersect(r.Hook.ExcludeTags, script.Tags) {
-		logSkip(file.Name(), "(SKIP BY TAGS)")
-		return
-	}
-
-	// Skip non-regular files (dirs, symlinks, sockets, etc.)
-	if !file.Mode().IsRegular() {
-		log.Debugf("[lefthook] file %s is not a regular file, skipping", file.Name())
-		return
-	}
-
-	// Make sure file is executable
-	if (file.Mode() & executableMask) == 0 {
-		if err := r.Fs.Chmod(path, executableFileMode); err != nil {
-			log.Errorf("Couldn't change file mode to make file executable: %s", err)
-			r.fail(file.Name(), "")
-			return
-		}
-	}
-
-	var args []string
-	if len(script.Runner) > 0 {
-		args = strings.Split(script.Runner, " ")
-	}
-
-	args = append(args, path)
-	args = append(args, r.GitArgs[:]...)
 
 	if script.Interactive && !r.DisableTTY && !r.Hook.Follow {
 		log.StopSpinner()
@@ -273,7 +247,7 @@ func (r *Runner) runCommands() {
 
 	for _, name := range commands {
 		if r.failed.Load() && r.Hook.Piped {
-			logSkip(name, "(SKIP BY BROKEN PIPE)")
+			logSkip(name, "broken pipe")
 			continue
 		}
 
@@ -297,7 +271,7 @@ func (r *Runner) runCommands() {
 
 	for _, name := range interactiveCommands {
 		if r.failed.Load() {
-			logSkip(name, "(SKIP INTERACTIVE BY FAILED)")
+			logSkip(name, "non-interactive commands failed")
 			continue
 		}
 
@@ -306,34 +280,9 @@ func (r *Runner) runCommands() {
 }
 
 func (r *Runner) runCommand(name string, command *config.Command) {
-	if command.Skip != nil && command.DoSkip(r.Repo.State()) {
-		logSkip(name, "(SKIP BY SETTINGS)")
-		return
-	}
-
-	if intersect(r.Hook.ExcludeTags, command.Tags) {
-		logSkip(name, "(SKIP BY TAGS)")
-		return
-	}
-
-	if intersect(r.Hook.ExcludeTags, []string{name}) {
-		logSkip(name, "(SKIP BY NAME)")
-		return
-	}
-
-	if err := command.Validate(); err != nil {
-		r.fail(name, "")
-		return
-	}
-
-	args, err := r.buildCommandArgs(command)
+	args, err := r.prepareCommand(name, command)
 	if err != nil {
-		log.Error(err)
-		logSkip(name, "(SKIP. ERROR)")
-		return
-	}
-	if len(args) == 0 {
-		logSkip(name, "(SKIP. NO FILES FOR INSPECTION)")
+		logSkip(name, err.Error())
 		return
 	}
 
@@ -352,114 +301,21 @@ func (r *Runner) runCommand(name string, command *config.Command) {
 	}, r.Hook.Follow)
 }
 
-func (r *Runner) buildCommandArgs(command *config.Command) ([]string, error) {
-	filesCommand := r.Hook.Files
-	if command.Files != "" {
-		filesCommand = command.Files
-	}
+func (r *Runner) preRun(opts ExecuteOptions) {
+	// if !config.HookHasStagedFiles(r.HookName) {
+	// 	return
+	// }
 
-	filesTypeToFn := map[string]func() ([]string, error){
-		config.SubStagedFiles: r.Repo.StagedFiles,
-		config.PushFiles:      r.Repo.PushFiles,
-		config.SubAllFiles:    r.Repo.AllFiles,
-		config.SubFiles: func() ([]string, error) {
-			return r.Repo.FilesByCommand(filesCommand)
-		},
-	}
-
-	runString := command.Run
-	for filesType, filesFn := range filesTypeToFn {
-		// Checking substitutions and skipping execution if it is empty.
-		//
-		// Special case - `files` option: return if the result of files
-		// command is empty.
-		if strings.Contains(runString, filesType) ||
-			filesCommand != "" && filesType == config.SubFiles {
-			files, err := filesFn()
-			if err != nil {
-				return nil, fmt.Errorf("error replacing %s: %s", filesType, err)
-			}
-			if len(files) == 0 {
-				return nil, nil
-			}
-
-			filesPrepared := prepareFiles(command, files)
-			if len(filesPrepared) == 0 {
-				return nil, nil
-			}
-
-			runString = replaceQuoted(runString, filesType, filesPrepared)
-		}
-	}
-
-	runString = strings.ReplaceAll(runString, "{0}", strings.Join(r.GitArgs, " "))
-	for i, gitArg := range r.GitArgs {
-		runString = strings.ReplaceAll(runString, fmt.Sprintf("{%d}", i+1), gitArg)
-	}
-
-	log.Debug("[lefthook] executing: ", runString)
-
-	return strings.Split(runString, " "), nil
-}
-
-func prepareFiles(command *config.Command, files []string) []string {
-	if files == nil {
-		return []string{}
-	}
-
-	log.Debug("[lefthook] files before filters:\n", files)
-
-	files = filterGlob(files, command.Glob)
-	files = filterExclude(files, command.Exclude)
-	files = filterRelative(files, command.Root)
-
-	log.Debug("[lefthook] files after filters:\n", files)
-
-	// Escape file names to prevent unexpected bugs
-	var filesEsc []string
-	for _, fileName := range files {
-		if len(fileName) > 0 {
-			filesEsc = append(filesEsc, shellescape.Quote(fileName))
-		}
-	}
-
-	log.Debug("[lefthook] files after escaping:\n", filesEsc)
-
-	return filesEsc
-}
-
-func replaceQuoted(source, substitution string, files []string) string {
-	for _, elem := range [][]string{
-		{"\"", "\"" + substitution + "\""},
-		{"'", "'" + substitution + "'"},
-		{"", substitution},
-	} {
-		quote := elem[0]
-		sub := elem[1]
-		if !strings.Contains(source, sub) {
-			continue
-		}
-
-		quotedFiles := files
-		if len(quote) != 0 {
-			quotedFiles = make([]string, 0, len(files))
-			for _, fileName := range files {
-				quotedFiles = append(quotedFiles,
-					quote+surroundingQuotesRegexp.ReplaceAllString(fileName, "$1")+quote)
-			}
-		}
-
-		source = strings.ReplaceAll(
-			source, sub, strings.Join(quotedFiles, " "),
-		)
-	}
-
-	return source
+	// Hide unstaged changes
+	// Save stash name
 }
 
 func (r *Runner) run(opts ExecuteOptions, follow bool) {
 	log.SetName(opts.name)
 	defer log.UnsetName(opts.name)
+
+	defer r.postRun(opts)
+	r.preRun(opts)
 
 	if (follow || opts.interactive) && !r.SkipSettings.SkipExecution() {
 		log.Info(log.Cyan("\n  EXECUTE > "), log.Bold(opts.name))
@@ -495,6 +351,16 @@ func (r *Runner) run(opts ExecuteOptions, follow bool) {
 	log.Infof("\n")
 }
 
+func (r *Runner) postRun(opts ExecuteOptions) {
+	// if !config.HookHasStagedFiles(r.HookName) {
+	// 	return
+	// }
+
+	// Restage files which were staged (if configured)
+	// Return if there is no stash set in preRun
+	// Unstash it
+}
+
 // Returns whether two arrays have at least one similar element.
 func intersect(a, b []string) bool {
 	intersections := make(map[string]struct{}, len(a))
@@ -513,5 +379,12 @@ func intersect(a, b []string) bool {
 }
 
 func logSkip(name, reason string) {
-	log.Info(fmt.Sprintf("%s: %s", log.Bold(name), log.Yellow(reason)))
+	log.Info(
+		fmt.Sprintf(
+			"%s: %s %s",
+			log.Bold(name),
+			log.Gray("(skip)"),
+			log.Yellow(reason),
+		),
+	)
 }
